+++ src/R_willowGro.cpp
@@ -117,7 +117,7 @@ SEXP willowGro(
     double *centcoefs = REAL(CENTCOEFS);
     int centTimestep = INTEGER(CENTTIMESTEP)[0];
     double *centks = REAL(CENTKS);
-    int soilLayers = INTEGER(SOILLAYERS)[0];
+    size_t soilLayers = INTEGER(SOILLAYERS)[0];
     double *soilDepths = REAL(SOILDEPTHS);
     double *cws = REAL(CWS);
     double hydrDist = INTEGER(HYDRDIST)[0];
@@ -747,7 +747,7 @@ SEXP willowGro(
 		results->respiration[i] = Resp / (24*centTimestep);
 		results->soil_evaporation[i] = soilEvap;
 		results->leaf_psim[i] = LeafPsim;
-		for(int layer = 0; layer < soilLayers; ++layer) {
+		for(size_t layer = 0; layer < soilLayers; ++layer) {
 			results->psim[layer + i * soilLayers] = psi[layer + i * soilLayers];
 			results->water_status[layer + i * soilLayers] = water_status[layer + i * soilLayers];
 			results->root_distribution[layer + i * soilLayers] = root_distribution[layer + i * soilLayers];
